// src/generated/api/orders/models/orderDetailDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */
import type { OrderDetailDtoStatus } from './orderDetailDtoStatus';
import type { OrderDetailDtoPaymentStatus } from './orderDetailDtoPaymentStatus';
import type { OrderDetailDtoFulfillmentStatus } from './orderDetailDtoFulfillmentStatus';
import type { OrderRecipientDto } from './orderRecipientDto';
import type { OrderItemDto } from './orderItemDto';
import type { OrderStatusHistoryDto } from './orderStatusHistoryDto';

export interface OrderDetailDto {
  /** @pattern ^[1-9][0-9]*$ */
  id: string;
  orderNo: string;
  status: OrderDetailDtoStatus;
  paymentStatus: OrderDetailDtoPaymentStatus;
  fulfillmentStatus: OrderDetailDtoFulfillmentStatus;
  paymentMethod: string;
  shippingMethod: string;
  /** @pattern ^[1-9][0-9]*$ */
  branchId: string;
  branchName: string;
  warehouseName: string;
  grandTotal: string;
  itemCount: number;
  recipient: OrderRecipientDto;
  placedAt: string;
  version: number;
  currencyCode: string;
  pricesIncludeTax: boolean;
  subtotal: string;
  discountTotal: string;
  shippingTotal: string;
  taxTotal: string;
  /** @nullable */
  customerNote?: string | null;
  items: OrderItemDto[];
  statusHistory: OrderStatusHistoryDto[];
}